refactor(cockpit): extract RailItem into sibling module (#1354)
Tenth wedge of the cockpit_ui.py god-module split tracked by #1354.

Moves RailItem (the rail-row ListItem widget — 243 LOC carrying the
CockpitItem render path, alert-severity CSS classes from #989, the
heartbeat/working/stuck glyph table, project-rollup/user-action
diamond precedence from #1390/#1520, and the wrapped alert subtitle
rendering) into a new sibling module src/pollypm/cockpit_rail_item.py
with a re-export shim left in cockpit_ui.py.

The alert-subtitle helpers (_wrap_alert_reason /
_rail_alert_subtitle_width) stay in cockpit_ui and are lazy-imported
from inside RailItem.update_body — same cycle-break pattern used by
_RollupItem (#1700) and _InboxListItem (#1727) for similar inbox-row
extractions. _strip_trailing_spark continues to be imported lazily
from cockpit_rail as before.

cockpit_ui.py: 18,543 -> 18,303 LOC (-240).
